Remove Duplicate code in Applicaiton Detail
