+++ frontend/src/ClubManage/ApplicantTab.js
@@ -4,7 +4,7 @@ import { connect } from "react-redux";
 import { withRouter } from "react-router";
 import { Card } from "react-bootstrap";
 import * as actionCreators from "../store/actions/index";
-import ApplicationDetail from "./ApplicationDetail";
+import ApplicationDetail from "../Apply/ApplicationDetail";
 
 class ApplicantTab extends Component {
   state = {};
